Refactor tests imports
