+++ zenml/conftest.py
@@ -21,12 +21,12 @@ from pathlib import Path
 import pytest
 
 import zenml
-from zenml.core.backends import BaseBackend
-from zenml.core.datasources import BaseDatasource
-from zenml.core.metadata import ZenMLMetadataStore
-from zenml.core.pipelines import BasePipeline
-from zenml.core.repo import Repository, ZenMLConfig
-from zenml.core.steps import BaseStep
+from zenml.backends import BaseBackend
+from zenml.datasources import BaseDatasource
+from zenml.metadata import ZenMLMetadataStore
+from zenml.pipelines import BasePipeline
+from zenml.repo import Repository, ZenMLConfig
+from zenml.steps import BaseStep
 from zenml.utils import path_utils
 
 # Nicholas a way to get to the root
